Refactoring - frontend filstruktur
