+++ web/src/frontend/src/app/index.test.tsx
@@ -1,8 +1,8 @@
 import * as React from "react";
 import * as ReactDOM from "react-dom";
-import IntlProvider from "../../intlProvider";
+import IntlProvider from "../intlProvider";
 
-import App from "./App";
+import App from "./";
 
 it("renders without crashing", () => {
 	const div = document.createElement("div");
