+++ web/src/frontend/src/app/index.tsx
@@ -1,14 +1,22 @@
 import * as React from "react";
-import "../styles/app.css";
+import "./styles/app.css";
 import { FormattedMessage } from "react-intl";
+import { Route, Switch } from "react-router";
+import Start from "./start";
+import Skjema from "./skjema";
 
 class App extends React.Component<{}, {}> {
 	render() {
+		// const { match } = this.props;
 		return (
 			<div className="app-digisos">
 				<h1>
 					<FormattedMessage id="skjema.tittel" />
 				</h1>
+				<Switch>
+					<Route path={`/`} exact component={Start} />
+					<Route path={`/skjema`} component={Skjema} />
+				</Switch>
 				{this.props.children}
 			</div>
 		);
