--- web/src/frontend/src/app/reducers/index.ts
@@ -1,9 +1,0 @@
-const defaultState = {
-	soknadId: undefined
-};
-
-const AppReducer = (state = defaultState, action = {}) => {
-	return state;
-};
-
-export default AppReducer;
