+++ web/src/frontend/src/app/skjema/components/knapperad/knapperad.less
