+++ web/src/frontend/src/app/skjema/konstanter.ts
