+++ web/src/frontend/src/app/skjema/steg/bosituasjon/index.tsx
@@ -1,0 +1,10 @@
+import * as React from "react";
+import Steg from "../../../../skjema/components/steg";
+
+class Bosted extends React.Component<{}, {}> {
+	render() {
+		return <Steg tittel="Bosituasjon" />;
+	}
+}
+
+export default Bosted;
