--- web/src/frontend/src/app/soknadsskjema/Steg2.tsx
@@ -1,9 +1,0 @@
-import * as React from "react";
-
-class Steg2 extends React.Component<{}, {}> {
-	render() {
-		return <div>Steg 2</div>;
-	}
-}
-
-export default Steg2;
