--- web/src/frontend/src/app/soknadsskjema/Steg6.tsx
@@ -1,9 +1,0 @@
-import * as React from "react";
-
-class Steg3 extends React.Component<{}, {}> {
-	render() {
-		return <div>Steg 6</div>;
-	}
-}
-
-export default Steg3;
