--- web/src/frontend/src/app/soknadsskjema/Steg7.tsx
@@ -1,9 +1,0 @@
-import * as React from "react";
-
-class Steg3 extends React.Component<{}, {}> {
-	render() {
-		return <div>Steg 7</div>;
-	}
-}
-
-export default Steg3;
