+++ web/src/frontend/src/redux/types.ts
