+++ web/src/frontend/src/skjema/components/bolk/bolk.less
