+++ web/src/frontend/src/skjema/components/steg/index.tsx
