+++ web/src/frontend/src/skjema/components/steg/steg.less
