+++ web/src/frontend/src/skjema/components/stegIndikator/index.tsx
