+++ web/src/frontend/src/skjema/faktum/components/FaktumInput.less
