+++ web/src/frontend/src/skjema/faktum/utils.ts
